+++ packages/central-server/src/database/models/index.js
@@ -27,3 +27,4 @@ export { SurveyScreenComponentModel as SurveyScreenComponent } from './SurveyScr
 export { SurveyScreenModel as SurveyScreen } from './SurveyScreen';
 export { UserModel as User } from './User';
 export { UserEntityPermissionModel as UserEntityPermission } from './UserEntityPermission';
+export { UserAccountModel as UserAccount } from './UserAccount';
